ai_video_workflow/requirements.txt
google-genai>=1.51.0
requests>=2.31.0
